Initial Unity Project Commit

+++ Assets/Scripts/CollectableRotate.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class CollectableRotate : MonoBehaviour
 {
     // Start is called once before the first execution of Update after the MonoBehaviour is created
-    [SerializeField] int rotateSpeed = 1;
+    [SerializeField] float rotateSpeed = 1;
     void Update()
     {
         transform.Rotate(0, rotateSpeed, 0, Space.World);
